Move 'get_latest_image_timestamp' and 'get_latest_folder' from scripts to src
